backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai>=1.0.0
googlemaps
